+++ src/auth/auth.controller.ts
@@ -20,7 +20,10 @@ export class AuthController {
   private readonly returnUrl = `http://${this.configService.get<string>('NEST_API_BASE_URL')}:${this.configService.get<string>('NEST_API_PORT')}`;
 
   @Get('login')
-  @ApiResponse({ status: 302, description: '성공적으로 스팀 로그인 URL을 찾아서 리다이렉트 합니다'})
+  @ApiResponse({
+    status: 302,
+    description: '성공적으로 스팀 로그인 URL을 찾아서 리다이렉트 합니다',
+  })
   async steamLogin(@Req() req: Request, @Res() res: Response) {
     const steamLoginUrl = await this.authService.getSteamLoginUrl();
     const referer = req.headers.referer;
@@ -32,8 +35,15 @@ export class AuthController {
   }
 
   @Get('login/return')
-  @ApiResponse({ status: 304, description: 'Steam OpenID 인증이 성공하여 로그인 이전 페이지로 돌아갑니다' })
-  @ApiResponse({ status: 401, description: 'Steam OpenID 인증 과정에서 오류가 발생하여 인증이 실패했습니다' })
+  @ApiResponse({
+    status: 304,
+    description: 'Steam OpenID 인증이 성공하여 로그인 이전 페이지로 돌아갑니다',
+  })
+  @ApiResponse({
+    status: 401,
+    description:
+      'Steam OpenID 인증 과정에서 오류가 발생하여 인증이 실패했습니다',
+  })
   async steamReturn(@Req() req: Request, @Res() res: Response) {
     try {
       const tokens = await this.authService.verifySteamResponse(req.url);
@@ -51,7 +61,10 @@ export class AuthController {
   }
 
   @Get('logout')
-  @ApiResponse({ status: 200, description: '모든 쿠키가 삭제되어 성공적으로 로그아웃 되었습니다' })
+  @ApiResponse({
+    status: 200,
+    description: '모든 쿠키가 삭제되어 성공적으로 로그아웃 되었습니다',
+  })
   async logout(@Res() res: Response) {
     this.authService.clearAllCookies(res);
     res.json({ success: true });
