+++ src/auth/auth.module.ts
@@ -5,10 +5,10 @@ import { HttpModule } from '@nestjs/axios';
 import { ConfigModule } from '@nestjs/config';
 import { JwtModule } from '@nestjs/jwt';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { RefreshToken } from './entities/refreshtoken.entity';
-import { UserService } from 'src/user/user.service';
-import { UserModule } from 'src/user/user.module';
-import { User } from 'src/user/entities/user.entity';
+import { RefreshToken } from '../entities/refreshtoken.entity';
+import { UserModule } from 'src/modules/user.module';
+import { User } from 'src/entities/user.entity';
+import { UserService } from 'src/services/user.service';
 
 @Module({
   imports: [
