--- src/auth/decorator/user.decorator.ts
@@ -1,11 +1,0 @@
-import { ExecutionContext, createParamDecorator } from '@nestjs/common';
-import { User } from 'src/user/entities/user.entity';
-
-export const UserDeco = createParamDecorator(
-  (data: keyof User | undefined, ctx: ExecutionContext) => {
-    const req = ctx.switchToHttp().getRequest();
-    const user = req.user;
-
-    return data ? user[data] : user;
-  },
-);
