+++ src/decorators/user.decorator.ts
@@ -1,0 +1,11 @@
+import { ExecutionContext, createParamDecorator } from '@nestjs/common';
+import { User } from 'src/entities/user.entity';
+
+export const UserDeco = createParamDecorator(
+  (data: keyof User | undefined, ctx: ExecutionContext) => {
+    const req = ctx.switchToHttp().getRequest();
+    const user = req.user;
+
+    return data ? user[data] : user;
+  },
+);
