--- src/game/game.module.ts
@@ -1,15 +1,0 @@
-import { Module } from '@nestjs/common';
-import { GameService } from './game.service';
-import { GameController } from './game.controller';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { Game } from './entities/game.entity';
-import { AuthModule } from 'src/auth/auth.module';
-import { UserModule } from 'src/user/user.module';
-
-@Module({
-  imports: [TypeOrmModule.forFeature([Game]), AuthModule, UserModule],
-  exports: [GameService],
-  controllers: [GameController],
-  providers: [GameService],
-})
-export class GameModule {}
