--- src/game/game.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { GameService } from './game.service';
-
-describe('GameService', () => {
-  let service: GameService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [GameService],
-    }).compile();
-
-    service = module.get<GameService>(GameService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
