+++ src/tests/todo.service.spec.ts
@@ -1,0 +1,18 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { TodoService } from '../services/todo.service';
+
+describe('TodoService', () => {
+  let service: TodoService;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      providers: [TodoService],
+    }).compile();
+
+    service = module.get<TodoService>(TodoService);
+  });
+
+  it('should be defined', () => {
+    expect(service).toBeDefined();
+  });
+});
