--- src/todo/todo.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { TodoService } from './todo.service';
-
-describe('TodoService', () => {
-  let service: TodoService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [TodoService],
-    }).compile();
-
-    service = module.get<TodoService>(TodoService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
