--- src/user/user.module.ts
@@ -1,19 +1,0 @@
-import { Module, forwardRef } from '@nestjs/common';
-import { UserService } from './user.service';
-import { UserController } from './user.controller';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { User } from './entities/user.entity';
-import { HttpModule } from '@nestjs/axios';
-import { AuthModule } from 'src/auth/auth.module';
-
-@Module({
-  imports: [
-    TypeOrmModule.forFeature([User]),
-    HttpModule,
-    forwardRef(() => AuthModule),
-  ],
-  exports: [UserService],
-  controllers: [UserController],
-  providers: [UserService],
-})
-export class UserModule {}
